Moved SamplingError to structs::error.rs
