+++ src/api.rs
@@ -1,10 +1,10 @@
+use crate::structs::SamplingError;
 use nalgebra::SVector;
 use serde::{Deserialize, Serialize};
 use std::io::Write;
 use std::io::{self, Read};
 use std::net;
 
-use crate::adherer_core::SamplingError;
 use crate::structs::Domain;
 use crate::{structs::Classifier, utils::svector_to_array};
 
@@ -96,10 +96,7 @@ impl<const N: usize> From<io::Error> for SamplingError<N> {
 }
 
 impl<const N: usize> Classifier<N> for RemoteClassifier<N> {
-    fn classify(
-        &mut self,
-        p: SVector<f64, N>,
-    ) -> Result<bool, crate::adherer_core::SamplingError<N>> {
+    fn classify(&mut self, p: SVector<f64, N>) -> Result<bool, SamplingError<N>> {
         if !self.domain.contains(&p) {
             return Err(SamplingError::OutOfBounds);
         }
